Minor scripts refactor
